fix: recover single- layout and improve receipt styles
